move landing data to data folder
