+++ src/data/Landing.js
@@ -1,0 +1,36 @@
+import item from "../../public/svg/item.svg";
+
+export const products = {
+  item1: {
+    title: "item",
+    image: item,
+    alt: "item",
+    stack: false,
+  },
+  item2: {
+    title: "Lorem ipsum dolor",
+    oldPrice: "$00.00",
+    newPrice: "$00.00",
+    description:
+      "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vestibulum convallis mi ut velit porttitor placerat. Nulla egestas mattis magna, ut luctus ligula mollis sit amet. Nam vitae leo dignissim, sollicitudin eros in.",
+    image: item,
+    alt: "item",
+    stack: true,
+  },
+  item3: {
+    title: "Lorem ipsum dolor",
+    oldPrice: "$00.00",
+    newPrice: "$00.00",
+    description:
+      "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vestibulum convallis mi ut velit porttitor placerat. Nulla egestas mattis magna, ut luctus ligula mollis sit amet. Nam vitae leo dignissim, sollicitudin eros in.",
+    image: item,
+    alt: "item",
+    stack: true,
+  },
+  item4: {
+    title: "item",
+    image: item,
+    alt: "item",
+    stack: false,
+  },
+};
